Error message dialog now works.
git-svn-id: svn://svn.code.sf.net/p/wxdsgn/code/trunk@1055 917957a4-63a1-4874-a357-9ccf276ff571

+++ packman2extended/DevPakInfo.h
@@ -39,7 +39,7 @@ public:
     wxArrayString TargetDirs,
     DestinationDirs,
     InstalledFiles;
-    int currentFileNumber;
+    size_t currentFileNumber;
     StatusType pakStatus;
 
 public:
